A crawl a user can actually run: `scrapex contractors`
REQ-24. He asked how a NEW USER would crawl the directory, and the measured
answer was that they could not — by any supported path:

  * `pyproject.toml` ships `include = ["scrapex*"]`, so `tools/` never reaches an
    installed user. `pip install` does not put the script on their machine.
  * `scrapex crawl` takes a `source_key from sources.yaml`, and muqawil is not in
    `sources.yaml`.
  * `cli.py` had ZERO references to `muqawil`, `partitioncrawl`, `snapshotcrawl`
    or `generic_record`.
  * `jobs.py` and the panel have zero references — the update button runs the
    price connectors.

THAT IS WHY HIS PANEL SAID "ENGINE NOT DETECTED" WHILE A CRAWL WAS RUNNING. The
crawl was `python tools/crawl_muqawil_listing.py`, a developer script writing
straight into SQLite. It never went near the Engine and did not need to.
Everything built for this directory — the provable partition, the sightings
ledger, the resume, the approval from disk — was reachable only by cloning the
repository.

So the implementation moved into the package as `scrapex/contractors.py`, with
`scrapex contractors --plan / --crawl / --approve / --coverage`.
`tools/crawl_muqawil_listing.py` is now a four-line pointer, kept because six
documents and one running command name that path, and a file answering "no such
file" teaches nothing.

ONE IMPLEMENTATION, TWO FRONT DOORS. `add_arguments` and `run` are shared by the
subcommand and by `python -m scrapex.contractors`, so neither can grow a flag the
other lacks — the rule `publish.workbook_tables` already follows. A test asserts
the flag set is declared in one place.

TWO IMPORT-TIME SIDE EFFECTS DELETED. The `sys.path.insert` existed only because
the file lived outside the package. The `sys.stdout.reconfigure` is unnecessary:
`cli.main` calls `_force_utf8_output()` before any subcommand runs, so `say`'s
guard is the belt and nothing has to be done at import. A library module that
reconfigures stdout when imported is a library module with an opinion about
someone else's process.

MEASURED BEFORE IMPORTING IT FROM `cli`: `scrapex.contractors` is ~1.7 s
cumulative to import and `scrapex.cli` is already ~1.74 s, because the heavy part
— `connectors.base` and `HttpFetcher` — is already in cli's graph via
`connectors.factory`. The marginal cost is its own ~3 ms self time, so a
module-level import is right and lazy-importing would have been complexity paid
for nothing.

THE TESTS MOVED WITH IT, which is what OP-28 was really about: they were guarding
a script no user could run. The `importlib.util.spec_from_file_location` load is
gone, and the test that asserted the path hack is replaced by one asserting the
module is inside the package — the trap it guarded is now handled by ordinary
imports.

Verified: exit code 2 on all four refusals through both front doors; the old
`tools/` path still works and prints the new usage; 20 tests pass.

WHAT THIS DOES NOT DO: there is still no path from the panel. `jobs.py` does not
know this crawl, so it cannot be started, paused or resumed from the product's
own interface. The plan item is split — the CLI half ticked, the panel half left
open.
